add planet rotations
2 rotations each
- 1 around the planet (set in create_scene())
- 1 on each planets own axis (set in each planet constructor)

rotations around the planet were set to be accurate to the solar system, but we can alter it to make it more aesthetically pleasing if we decide.

individual planet-axis rotations are set but still need to be fine-tuned to be  accurate

I also changes the positions of the planets a bit, the z axis becomes irrelevant when rotating on the x-axis.

all rotations use a new rotations() class that offers more customization that the included one.